requirements.txt
dateutils==0.6.6
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-dateutil==2.5.3
pytz==2016.4
requests==2.9.1
six==1.10.0
Werkzeug==0.11.9
